feat: add custom user identifier support for throttle configuration and update documentation
