+++ packages/socketio-decorator/src/Models/ThrottleConfig.ts
@@ -1,4 +1,5 @@
 import { IThrottleStorage } from "../Interfaces/IThrottleStorage"
+import { Socket } from "socket.io"
 
 /**
  * Defines the configuration for event throttling.
@@ -25,4 +26,10 @@ export type ThrottleConfig = {
 	 * The throttle data store to use. By default, it uses in-memory storage.
 	 */
 	store?: new() => IThrottleStorage
+	/**
+	 * Function to get a custom identifier for the user, used to identify and retrieve user throttle data. If not provided, socket.id will be used.
+	 * @param {Socket} socket The socket instance
+	 * @returns {string | Promise<string>} A unique identifier for the user
+	 */
+	getUserIdentifier?: (socket: Socket) => string | Promise<string>
 }
