+++ packages/socketio-decorator/src/Wrappers/throttle/ThrottleManager.ts
@@ -3,6 +3,7 @@ import { IThrottleStorage } from "../../Interfaces/IThrottleStorage"
 import { IoCContainer } from "../../IoCContainer"
 import { SiodThrottleError } from "../../Models/Errors/SiodThrottleError"
 import { InMemoryThrottleStorage } from "./InMemoryThrottleStorage"
+import { Socket } from "socket.io"
 
 /**
  * Manages throttling state and checks for all socket events
@@ -24,18 +25,19 @@ export class ThrottleManager {
 
 	/**
 	 * Check if a request should be throttled
-	 * @param {string} clientId The client ID
+	 * @param {Socket} socket The socket instance
 	 * @param {string} eventName The event name
 	 * @param {number} limit Maximum requests allowed
 	 * @param {number} windowMs Time window in milliseconds
 	 * @returns {Promise<void>} Promise that resolves if the request is allowed, rejects with ThrottleError if throttled
 	 */
 	public static async checkThrottle (
-		clientId: string,
+		socket: Socket,
 		eventName: string,
 		limit: number,
 		windowMs: number
 	): Promise<void> {
+		const clientId = await this.getUserIdentifier(socket)
 		const now = Date.now()
 
 		const eventData = await this.store.get(clientId, eventName)
@@ -64,5 +66,17 @@ export class ThrottleManager {
 	 */
 	public static startPeriodicCleanup (interval: number = 3600000): NodeJS.Timeout {
 		return setInterval(() => this.store.cleanup(), interval)
+	}
+
+	/**
+	 * Get user identifier from socket
+	 * @param {Socket} socket The socket instance
+	 * @returns {Promise<string>} The user identifier
+	 */
+	private static async getUserIdentifier (socket: Socket): Promise<string> {
+		if (config.throttleConfig?.getUserIdentifier) {
+			return await Promise.resolve(config.throttleConfig.getUserIdentifier(socket))
+		}
+		return socket.id
 	}
 }
